size_t changed

+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: ijmari <ijmari@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/04 09:22:05 by ijmari            #+#    #+#             */
-/*   Updated: 2021/11/04 11:04:26 by ijmari           ###   ########.fr       */
+/*   Updated: 2021/11/11 20:40:09 by ijmari           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int			i;
+	size_t					i;
 	unsigned char			*t1;
 	unsigned char			*t2;
